add some typesafety now that we have discord.py-stubs, im too tired to finish today

+++ cmds/cmd_administration.py
@@ -77,6 +77,8 @@ class gdpr_functions:
         self.commands = {"delete": self.delete, "download": self.download}
 
     async def delete(self, message: discord.Message, guild_id: int, ramfs: Any, kramfs: Any) -> None:
+        if not message.guild:
+            return
 
         with db_hlapi(message.guild.id) as database:
             database.delete_guild_db()
@@ -130,6 +132,8 @@ Or if discord experienced a lag spike, consider retrying as the network may have
 
 
 async def gdpr_database(message: discord.Message, args: List[str], client: discord.Client, **kwargs: Any) -> Any:
+    if not message.guild:
+        return 1
 
     ramfs = kwargs["ramfs"]
 
@@ -158,6 +162,8 @@ async def gdpr_database(message: discord.Message, args: List[str], client: disco
 
 
 async def set_view_infractions(message: discord.Message, args: List[str], client: discord.Client, **kwargs: Any) -> Any:
+    if not message.guild:
+        return 1
 
     if args:
         gate = parse_boolean(args[0])
@@ -171,6 +177,8 @@ async def set_view_infractions(message: discord.Message, args: List[str], client
 
 
 async def set_prefix(message: discord.Message, args: List[str], client: discord.Client, **kwargs: Any) -> Any:
+    if not message.guild:
+        return 1
 
     if args:
         prefix = args[0]
@@ -303,4 +311,4 @@ commands = {
         }
     }
 
-version_info: str = "1.2.5"
+version_info: str = "pre2.0.0-DEV"
